Fix CR comments
